ISSUE-36 - Organize Environment and Logger init
- Organizing Logger
